+++ Sources/Tiles.swift
@@ -2,7 +2,7 @@ import Foundation
 import Archivable
 
 struct Tiles: Storable {
-    private let items: [UInt8 : [UInt32 : [UInt32 : UInt64]]]
+    private let items: [UInt8 : [UInt32 : [UInt32 : UInt32]]]
     
     var data: Data {
         .init()
@@ -11,13 +11,13 @@ struct Tiles: Storable {
             .flatMap { z in
                 Data()
                     .adding(z.key)
-                    .adding(UInt16(z.value.count))
+                    .adding(UInt8(z.value.count))
                     .adding(z
                         .value
                         .flatMap { x in
                             Data()
                                 .adding(x.key)
-                                .adding(UInt16(x.value.count))
+                                .adding(UInt8(x.value.count))
                                 .adding(x
                                     .value
                                     .flatMap { y in
@@ -32,9 +32,9 @@ struct Tiles: Storable {
     init(data: inout Data) {
         items = (0 ..< .init(data.number() as UInt8))
             .reduce(into: [:]) { z, _ in
-                z[data.number()] = (0 ..< .init(data.number() as UInt16))
+                z[data.number()] = (0 ..< .init(data.number() as UInt8))
                     .reduce(into: [:]) { x, _ in
-                        x[data.number()] = (0 ..< .init(data.number() as UInt16))
+                        x[data.number()] = (0 ..< .init(data.number() as UInt8))
                             .reduce(into: [:]) { y, _ in
                                 y[data.number()] = data.number()
                             }
@@ -42,11 +42,11 @@ struct Tiles: Storable {
             }
     }
     
-    init(items: [UInt8 : [UInt32 : [UInt32 : UInt64]]]) {
+    init(items: [UInt8 : [UInt32 : [UInt32 : UInt32]]]) {
         self.items = items
     }
     
-    subscript(x: Int, y: Int, z: Int) -> UInt64? {
+    subscript(x: Int, y: Int, z: Int) -> UInt32? {
         items[.init(z)]?[.init(x)]?[.init(y)]
     }
 }
